Moving engine subdirs into lib/
