+++ src/lib/license-matching/similarity-scorer.ts
