+++ src/lib/log/deal-generator.ts
@@ -1,6 +1,6 @@
 import { LogWriteStream } from "../data/file.js";
-import { Action } from "../engine/deal-generator/actions.js";
-import { DealRelevantEvent } from "../engine/deal-generator/events.js";
+import { Action } from "../deal-generator/actions.js";
+import { DealRelevantEvent } from "../deal-generator/events.js";
 import { DealStage } from '../hubspot/interfaces.js';
 import { License } from "../model/license.js";
 import { Transaction } from "../model/transaction.js";
