+++ src/tests/deal-generator/utils.ts
@@ -1,13 +1,13 @@
 import Chance from 'chance';
-import { ContactGenerator } from '../../lib/engine/contact-generator';
-import { updateContactsBasedOnMatchResults } from '../../lib/engine/contact-generator/update-contacts';
-import { DealGenerator } from '../../lib/engine/deal-generator';
-import { Action } from "../../lib/engine/deal-generator/actions";
-import { DealRelevantEvent } from '../../lib/engine/deal-generator/events';
+import { ContactGenerator } from '../../lib/contact-generator';
+import { updateContactsBasedOnMatchResults } from '../../lib/contact-generator/update-contacts';
+import { DealGenerator } from '../../lib/deal-generator';
+import { Action } from "../../lib/deal-generator/actions";
+import { DealRelevantEvent } from '../../lib/deal-generator/events';
 import { Engine } from '../../lib/engine/engine';
-import { RelatedLicenseSet } from '../../lib/engine/license-matching/license-grouper';
 import { Hubspot } from '../../lib/hubspot';
 import { DealStage } from '../../lib/hubspot/interfaces';
+import { RelatedLicenseSet } from '../../lib/license-matching/license-grouper';
 import { DealData } from "../../lib/model/deal";
 import { License, LicenseData } from "../../lib/model/license";
 import { ContactInfo } from '../../lib/model/record';
